+++ CHANGELOG.md
@@ -3,4 +3,4 @@ most recent version is listed first.
 
 
 ## **version:** v0.1.1
-- first release
+- make task buffers be thread safe: https://github.com/komuw/wijisqs/pull/11